+++ js/DESCRIPTION.js
@@ -1,0 +1,10 @@
+export const DESCRIPTION = [
+  'Интересное фото',
+  'Природа',
+  'Море',
+  'Деревья',
+  'Небо',
+  'Животные',
+  'Закат',
+  'Необычное фото',
+];
