+++ js/createComment.js
@@ -1,0 +1,3 @@
+export const createComment = () => [
+  'Комментарий',
+];
